+++ src/demo/Demo.vue
@@ -9,11 +9,12 @@ const TvBreadcrumbs = defineAsyncComponent(/* webpackChunkName: "TvBreadcrumbs" 
   <TvDemo
     :component="TvBreadcrumbs"
     :variants="demos"
+    :manual-emits="['item-click', 'navigate']"
     hide-background
     component-name="TvBreadcrumbs"
     npm-install="@todovue/tv-breadcrumbs"
     source-link="https://github.com/TODOvue/tv-breadcrumbs"
     url-clone="https://github.com/TODOvue/tv-breadcrumbs.git"
-    version="1.1.2"
+    version="1.1.3"
   />
 </template>